Session refactoring

+++ addon/sec_chan/internal/server/create.go
@@ -81,7 +81,7 @@ func (s impl) Create(ctx context.Context, req *proto.Create_Request) (res *proto
 	}
 
 	// Create a new session.
-	lifetime := uint16(req.GetLifetime())
+	lifetime := req.GetLifetime()
 	sess := session.NewSession(lifetime)
 	dic := netCtx.Response().Meta().Dictionary().(*net.HttpDictionary)
 	dic.SessionId = sess.Id()
